refactor: consolidate storage backends (14 files deleted)
async_sqlite_* (11→4):
- Merged 7 mixin/helper files into async_sqlite.py: support, runtime,
  transactions, connections, derived_stats, derived_actions, derived_products

schema_ddl_* (9→7):
- Merged schema_ddl_messages.py into schema_ddl_archive.py
- Merged schema_ddl_products.py inline into schema_ddl.py

query_store_* (9→5):
- Merged 4 tiny product query stores into query_store.py: product_status,
  product_actions, product_threads, product_summaries

queries/mappers_support.py merged into queries/mappers.py

All 4,320 tests pass.
