+++ polylogue/storage/backends/queries/mappers.py
@@ -1,8 +1,45 @@
-"""Small public root for row-mapper families."""
+"""Public root for row-mapper families and shared helpers."""
 
 from __future__ import annotations
 
-from polylogue.storage.backends.queries.mappers_archive import (
+import json
+import sqlite3
+from typing import Any
+
+from polylogue.errors import DatabaseError
+
+# ---------------------------------------------------------------------------
+# Shared row-mapper support helpers (formerly mappers_support.py)
+# These MUST be defined before the sub-module imports below, since the
+# sub-modules import _parse_json / _row_get from this module.
+# ---------------------------------------------------------------------------
+
+
+def _parse_json(raw: str | None, *, field: str = "", record_id: str = "") -> Any:
+    """Parse a JSON string with diagnostic context on failure."""
+    if not raw:
+        return None
+    try:
+        return json.loads(raw)
+    except json.JSONDecodeError as exc:
+        raise DatabaseError(
+            f"Corrupt JSON in {field} for {record_id}: {exc} (value starts: {raw[:80]!r})"
+        ) from exc
+
+
+def _row_get(row: sqlite3.Row, key: str, default: Any = None) -> Any:
+    """Get a column value, returning default if the column doesn't exist."""
+    try:
+        return row[key]
+    except (KeyError, IndexError):
+        return default
+
+
+# ---------------------------------------------------------------------------
+# Re-exports from sub-modules
+# ---------------------------------------------------------------------------
+
+from polylogue.storage.backends.queries.mappers_archive import (  # noqa: E402
     _row_to_action_event,
     _row_to_artifact_observation,
     _row_to_content_block,
@@ -10,7 +47,7 @@ from polylogue.storage.backends.queries.mappers_archive import (
     _row_to_message,
     _row_to_raw_conversation,
 )
-from polylogue.storage.backends.queries.mappers_products import (
+from polylogue.storage.backends.queries.mappers_products import (  # noqa: E402
     _row_to_day_session_summary_record,
     _row_to_session_phase_record,
     _row_to_session_profile_record,
@@ -18,7 +55,6 @@ from polylogue.storage.backends.queries.mappers_products import (
     _row_to_session_work_event_record,
     _row_to_work_thread_record,
 )
-from polylogue.storage.backends.queries.mappers_support import _parse_json, _row_get
 
 __all__ = [
     "_parse_json",
