+++ polylogue/storage/backends/schema_ddl.py
@@ -12,6 +12,9 @@ from polylogue.storage.backends.schema_ddl_archive import (
     ARCHIVE_STORAGE_DDL as _ARCHIVE_STORAGE_DDL,
 )
 from polylogue.storage.backends.schema_ddl_archive import (
+    MESSAGE_FTS_DDL as _MESSAGE_FTS_DDL,
+)
+from polylogue.storage.backends.schema_ddl_archive import (
     RAW_ARCHIVE_DDL as _RAW_ARCHIVE_DDL,
 )
 from polylogue.storage.backends.schema_ddl_aux import (
@@ -23,11 +26,14 @@ from polylogue.storage.backends.schema_ddl_aux import (
 from polylogue.storage.backends.schema_ddl_aux import (
     VEC0_DDL as _VEC0_DDL,
 )
-from polylogue.storage.backends.schema_ddl_messages import (
-    MESSAGE_FTS_DDL as _MESSAGE_FTS_DDL,
+from polylogue.storage.backends.schema_ddl_product_aggregates import (
+    SESSION_PRODUCT_AGGREGATE_DDL as _SESSION_PRODUCT_AGGREGATE_DDL,
 )
-from polylogue.storage.backends.schema_ddl_products import (
-    SESSION_PRODUCT_DDL as _SESSION_PRODUCT_DDL,
+from polylogue.storage.backends.schema_ddl_product_profiles import (
+    SESSION_PRODUCT_PROFILE_DDL as _SESSION_PRODUCT_PROFILE_DDL,
+)
+from polylogue.storage.backends.schema_ddl_product_timelines import (
+    SESSION_PRODUCT_TIMELINE_DDL as _SESSION_PRODUCT_TIMELINE_DDL,
 )
 
 SCHEMA_VERSION = 1
@@ -48,6 +54,13 @@ SCHEMA_DDL = (
 
 SCHEMA_DDL += _ACTION_EVENT_DDL
 SCHEMA_DDL += _ACTION_FTS_DDL
+
+_SESSION_PRODUCT_DDL = (
+    _SESSION_PRODUCT_PROFILE_DDL
+    + _SESSION_PRODUCT_TIMELINE_DDL
+    + _SESSION_PRODUCT_AGGREGATE_DDL
+)
+
 SCHEMA_DDL += _SESSION_PRODUCT_DDL
 
 __all__ = [
